+++ dmr/security/jwt/auth.py
@@ -324,7 +324,7 @@ class JWTAsyncAuth(_BaseJWTAuth, AsyncAuth):
         token: JWToken,
     ) -> None:
         """Set current user as authed for this request."""
-        set_request_attrs(request, user, token=token, include_auser=True)
+        set_request_attrs(request, user, token=token)
 
 
 @overload
@@ -361,17 +361,17 @@ def set_request_attrs(
     user: 'AbstractBaseUser',
     *,
     token: JWToken | None = None,
-    include_auser: bool = False,
 ) -> None:
     """Set all required properties to the authed request."""
     request.user = user
 
-    if include_auser:
+    # This is needed even for sync views for consistency,
+    # and wild `async_to_sync` / `sync_to_async` use-cases:
 
-        async def auser() -> 'AbstractBaseUser':  # noqa: WPS430
-            return user
+    async def auser() -> 'AbstractBaseUser':  # noqa: WPS430
+        return user
 
-        request.auser = auser
+    request.auser = auser
 
     if token is not None:
         request.__dmr_jwt__ = token  # type: ignore[attr-defined]