+++ CHANGELOG.md
@@ -39,6 +39,7 @@ of requirements for an API to count as public.
 - Fixed `datetime` validation when using `.to_response`, #938
 - Fixed a bug that `ObtainTokensAsyncController` was not setting
   the `request.auser` attribute, #953
+- Fixed a bug that `JWTSyncAuth` was not setting `request.auser`, #953
 
 ### Misc